Validate user attribute key
Creating a local copy of the user attribute key to avoid the original key going out of scope and becoming nil. This may happen when trying to set a user attribute prior to starting the SDK and key goes out of scope.

In addition to the local copy, further validation was added in both the setUserAttribute method and the method to validate the parameters to be used in setting a user attribute.