Update RequestBuilderTest to run in isolation
Prior to this, MPURLRequestBuilderTests required that MPKitContainerTests
ran first, otherwise the MPKitContainer singleton would not have kit 314,
and testURLRequestComposition would fail. This causes sporadic travis failures.

This could have also been addressed by removing the check for 314 from
testURLRequestComposition, but figured it's better to keep the test as is,
which checks for multiple kit codes.