Timing of logged events
Events (both app events and commerce events) now have a timestamp property, which gets populated automatically by the SDK an event is logged prior to the SDK being fully initialized. If set, this property will override the timestamp of messages when they are about to be persisted.

There was also a change in the fetching statement. Messages for an upload batch are now retrieved by timestamp, rather than the order they were inserted in the database.