Queueing events in the App Notification Handler
The app notification handler now takes advantage of the forwarding queue mechanism. If kits have not been initialized yet (config not received from server), the data will be held in a queue and once the configuration has been received and kits initialized, the forward queue items are replayed and the newly initialized kits will receive the data.