Remove NSDictionary generic specifications for checkForDeferredDeepLinkWithCompletionHandler callback
Because the callback for checkForDeferredDeepLinkWithCompletionHandler traces its NSDictionary all the way back to `[Branch getLatestReferringParams]`, which doesn’t specify generics for NSDictionary, there is no guarantee that the keys and values will be of type `<NSString *, NSString *>`. This proves to be an issue when bridging to Swift, as there are times when `[Branch getLatestreferringParams]` _can_ return a dictionary of type `[NSObject : AnyObject]`, which causes a crash at runtime because Swift is expecting `[String : String]`, but is getting `[String : NSNumber]`.

Therefore, the callback can only be as specific as its upstream dependencies, which is not at all specific. Thus, removing the <NSString *, NSString *> generic specification on the callback.

This bridges to `[NSObject : AnyObject]` when consumed in Swift, which requires the consumer to cast from `AnyObject` to the expected type (`String` in this case) when pulling values out of the dictionary.

Here is how I’ve implemented the contract change in our code:
```
if let canonicalURL = linkInfo?["$canonical_url"] as? String, let url = NSURL(string: canonicalURL) {
    navigateToLink(url)
}
```