Add option to control automatic session tracking
This commit adds a new option to the mParticle options object that
will allow automatic tracking of sessions to be turned off. Modifying
the code to respect this option will be done in a separate commit.

This commit also adds a corresponding property to the state machine and
adds unit tests for MParticleOptions