Allow session_id to be nullable
This commit allows session_id to be nullable in the messages and
uploads tables. It updates the create statements to remove the
foreign key constraint on session id and remove "NOT NULL" from
the column properties.

It also adds checks to insert null when writing messages and uploads
to the DB, in migration and in normal persistence methods.

The equality and copy methods have been updated as well as the
property declarations, since the session id property is now
nullable and a pointer instead of an int64.

Finally, this commit removes some conformance with NSCoding that is
unused and would otherwise need to be updated.