Implement startKitsAsync initialization option
This commit introduces an option to the SDK's initialization API to load partner-SDKs
asynchronously. This is already how partner-SDKs are started on first-app launch, due to
the requirement to download mParticle SDK configuration - so it should generally be safe.
Most SDK documentation states that they must be initialized synchronously, though in practice
this is false. The startKitsAsync options puts the app developer in control.