Use a single GCD queue for most SDK processing
This commit gets rid of the backend and persistence queues and
introduces a new message queue.

Except in a few situations (e.g. did enter background), the SDK
will uniformly dispatch to the message queue before doing any work.

It will also generally continue to run on the message queue (blocking
for I/O as needed) until the work is complete. This differs from
current SDK behavior where the SDK will get on and off of various
queues within the processing of a single incoming request.

When something needs to be run on the main thread (e.g. calls to
kits), we dispatch sync to the main thread so that we don't lose
our thread of execution and start executing the next message queue
task before we are done with this one.

Related to the above, this commit reduces the complexity caused by
large numbers of completion blocks. It removes the handlers for DB
and Network operations, opting to synchronously wait for results
since we're already on a background thread.

Finally, as a result of the changes already described, the unit tests
needed to be updated. They have been changed to dispatch to the
message queue before calling the backend controller or kit container.