fix: Transaction attributes revenue logic differs from Android
* Fix transaction attributes revenue logic differs from Android

In the Android SDK, the builder pattern used always ensures that product events have an instantiated transaction attributes object, while on iOS that was not done. This change simply ensures that the object always exists, as the actual revenue calculation logic appears to have been correct in this case. The problem was that it was trying to set the revenue value on a nil object.

* Add unit test for revenue value when commerce event doesn’t have transaction attributes set