refactor: Remove PLCrashReporter, MPExceptionHandler, and MP_CRASH_REPORTER
* First pass at refactoring out MPExceptionHandler

I confirmed that crash logs are still sent and are received correctly in the Live Stream. Also, the `MP_CRASH_REPORTER` feature flag has been removed entirely. The client now must instantiate the PLCrashReporter instance themselves as was already demonstrated in the included `Example` project's `AppDelegate` class.

NOTE: We are no longer providing handling for uncaught exceptions, the client is now responsible for that if they want it.

* Removed all references to PLCrashReporter

Clients using PLCrashReporter should use the new `- [MParticle logCrash:stackTrace:plCrashReport:]` method to include PLCrashReporter crash reports (as is already shown in the included Example project code). The `logException` methods will not include that specific key, but will still include stack trace and other exception information.

* Fixed non-deterministic unit test causing random failures

This resolves an issue with the `-[MPBackendControllerTests messageWithType:andLength:]` method used in a few unit tests causing the `testCrashReportBatchAndMessageByteLimits` test to randomly fail depending on things like CPU and memory usage fluctuations.

Previously we used the `MPMessageBuilder` to build a new message after truncating `longString` to the correct length. However, the other data included in the message is not deterministic as it includes things like CPU and memory usage which fluctuate during the test and change the message length. This could cause cases where the final message was larger than expected if any of those other values became longer (e.g. 9% CPU usage went to 10% CPU usage, adding another byte to the JSON).

To keep things deterministic, we are now directly modifying the contents of the original message data with the truncated string instead as that ensures that the final message data size is the size we intended for the test.