fix: update the x-mp-kits header value
This field was previously sending a list of kits that were activley receiving events. Now it will send a list of kits that have been configured, even if they are not activley receiving events due to the SDK state. For Example, a kit that was shut down due to the current user's logged-in status not matching its configurations directive previous would not have been included in this list but now will be included