fix: property memory annotations (#146)
This commit does 2 things.

First, it fixes a reported crash due to the MPAliasResponse NSString *errorResponse being marked as assign intead of copy (or strong), as well as fixing another potential instance of this that I found in MParticleOptions NSString *defaultAgent being marked as unsafe_unretained instead of copy (or strong). I changed both instances of this to copy.

Second, it removes all uses of the assign and unsafe_unretained memory annotations on value types, as that is the only valid option (strong and weak cannot be used for value types) and is the default anyway. Looking at Apple’s UIView.h for example, they omit them as well.