fix: update active kits header (#147)
The `x-mp-kits` header field was previously sending a list of kits that were actively receiving events. Now it will send a list of kits that have been configured, even if they are not actively receiving events due to the SDK state. For example, a kit that was shut down due to the current user's logged-in status not matching its configurations directive previous would not have been included in this list, but now will be included.